API

Macros

RestClientScaffold.@endpoint — Macro

@endpoint struct ... end

Define a struct that serves as an API endpoint.

This macro serves as a shorthand way of defining the basic endpoint methods:

pagename

parameters (if needed)

responsetype

payload (optionally)

Definitions for these methods are generated based on a non-standard endpoint declaration line at the top of the struct definition. This line takes the forms

"page/path?params..." -> ResultType           # Get request
input -> "page/path?params..." -> ResultType  # Post request

Both the page path and the parameters can contain references to global variables or fields of the struct, surrounded by curly braces. For convenience, a parameter by the same name as the field can be referred to by the field name alone (e.g. ?{var} instead of ?var={var}). The ResultType can be any type, but is typically a struct defined with @jsondef.

"page/{somefield}?param={another}&{globalvar}=7" -> ResultType

In more complex cases, arbitrary Julia code can be included in the curly braces. This code will be evaluated with the endpoint value bound the the anaphoric variable self. You can also reference type parameters of the struct.

"page/{if self.new "new" else "fetch" end}/{id}" -> ResultType

If that's not enough, you can also use an arbitrary expression in place of the endpoint string:

if self.create "pages/create/" else "pages/fetch/id/" * self.id end -> ResultType

When using an input -> page -> ResultType form for a post request, input should follow the form of a curly brace interpolation — be either a field name, global variable, or expression. It is used to define payload for the endpoint.

Examples

@endpoint struct ShuffleEndpoint <: SingleEndpoint
    "deck/{deck}/shuffle?remaining={ifelse(self.remaining, '1', '0')}" -> Deck
    deck::String
    remaining::Bool
end

This is equivalent to defining the struct by itself, and then separately defining the three basic endpoint methods.

struct ShuffleEndpoint <: SingleEndpoint
    deck::String
    remaining::Bool
end

RestClientScaffold.pagename(shuf::ShuffleEndpoint) =
    "deck/$(shuf.deck)/shuffle"
RestClientScaffold.parameters(shuf::ShuffleEndpoint) =
    ["remaining" => string(shuf.remaining)]
RestClientScaffold.responsetype(shuf::ShuffleEndpoint) = Deck

RestClientScaffold.@jsondef — Macro

@jsondef [kind] struct ... end

Define a struct that can be used with JSON3.

This macro conveniently combines the following pieces:

@kwdef to define keyword constructors for the struct.

Custom Base.show method to show the struct with keyword arguments, omitting default values.

StructTypes.StructType to define the struct type, and StructTypes.names (if needed) to define the JSON field mapping.

RestClientScaffold.dataformat to declare that this struct is JSON-formatted.

Optionally the JSON representation kind can be specified. It defaults to Struct, but can any of: Struct, Dict, Array, Vector, String, Number, Bool, Nothing.

Soft JSON3 dependency

This macro is implemented in a package extension, and so requires JSON3 to be loaded before it can be used.

Examples

@jsondef struct DocumentStatus
    exists::Bool  # Required, goes by 'exists' in JSON too
    status:"document_status"::Union{String, Nothing} = nothing
    url:"document_url"::Union{String, Nothing} = nothing
    age::Int = 0  # Known as 'age' in JSON too, defaults to 0
end

Request types

RestClientScaffold.Request — Type

Request{E<:AbstractEndpoint}

A request to an API endpoint, with a specific configuration.

This is the complete set of information required to make a request to an API.

See also: AbstractEndpoint, RequestConfig.

Data flow

         ╭─╴config╶────────────────────────────╮
         │     ╎                               │
         │     ╎        ╭─▶ responsetype ╾─────┼────────────────┬──▶ dataformat ╾───╮
Request╶─┤     ╰╶╶╶╶╶╶╶╶│                      │                ╰─────────╮         │
         │              ├─▶ pagename ╾───╮     │      ┌╶╶╶ debug╴╴╴┐      │  ╭──────╯
         │              │                ├──▶ url ╾─┬─━─▶ request ╾━┬─▶ interpret ╾──▶ data
         ├─╴endpoint╶───┼─▶ parameters ╾─╯          │               │                   │
         │              │                           │               │             ╭─────╯
         │              ├─▶ parameters ╾────────────┤               ╰─────────╮   │
         │              │                           │                      postprocess ╾──▶ result
         │             *╰─▶ payload ─▶ writepayload╶╯                           │
         │                    ╰─▶ dataformat ╾╯                                 │
         ╰─────────┬────────────────────────────────────────────────────────────╯
                   ╰────▶ validate (before initiating the request)

 * Only for POST requests   ╶╶ Optional first argument

RestClientScaffold.RequestConfig — Type

RequestConfig

The general configuration for a request to the API, not tied to any specific endpoint.

RestClientScaffold.AbstractEndpoint — Type

AbstractEndpoint

Abstract supertype for API endpoints.

Usually you will want to subtype either SingleEndpoint or ListEndpoint, which share the same interface as AbstractEndpoint but have additional semantics.

Interface

pagename([config::RequestConfig], endpoint::AbstractEndpoint) -> String
headers([config::RequestConfig], endpoint::AbstractEndpoint) -> Vector{Pair{String, String}}
parameters([config::RequestConfig], endpoint::AbstractEndpoint) -> Vector{Pair{String, String}}
responsetype(endpoint::AbstractEndpoint) -> Union{Type, Nothing}
validate([config::RequestConfig], endpoint::AbstractEndpoint) -> Bool
postprocess([response::Downloads.Response], request::Request, data) -> Any

All of these functions but pagename have default implementations.

See also: Request, dataformat, interpretresponse.

RestClientScaffold.SingleEndpoint — Type

SingleEndpoint <: AbstractEndpoint

Abstract supertype for API endpoints that return a single value.

See also: AbstractEndpoint, SingleResponse, Single.

RestClientScaffold.ListEndpoint — Type

ListEndpoint <: AbstractEndpoint

Abstract supertype for API endpoints that return a list of values.

See also: AbstractEndpoint, ListResponse, List.

Response types

RestClientScaffold.SingleResponse — Type

SingleResponse{T}

Abstract supertype for responses that contain a single T item and (optionally) metadata.

Interface

Subtypes of SingleResponse may need to define these two methods:

contents(single::SingleResponse{T}) -> T
metadata(single::SingleResponse{T}) -> Dict{Symbol, Any}

Both have generic implementations that are sufficient for simple cases.

RestClientScaffold.Single — Type

Single{T, E<:SingleEndpoint}

Holds a single value of type T returned from an API endpoint, along with request information and metadata.

See also: SingleEndpoint, SingleResponse.

RestClientScaffold.ListResponse — Type

ListResponse{T}

Abstract supertype for responses that contain a list of T items and (optionally) metadata.

Interface

Subtypes of ListResponse may need to define these two methods:

contents(list::ListResponse{T}) -> Vector{T}
metadata(list::ListResponse{T}) -> Dict{Symbol, Any}

Both have generic implementations that are sufficient for simple cases.

RestClientScaffold.List — Type

List{T, E<:ListEndpoint}

Holds a list of values of type T returned from an API endpoint, along with request information and metadata.

See also: ListEndpoint, ListResponse.

Request interface

RestClientScaffold.pagename — Function

pagename([config::RequestConfig], endpoint::AbstractEndpoint) -> String

Return the name of the page for the given endpoint.

This is combined with the base URL and parameters to form the full URL for the request.

Note

Part of the AbstractEndpoint interface.

RestClientScaffold.headers — Function

headers([config::RequestConfig], endpoint::AbstractEndpoint) -> Vector{Pair{String, String}}

Return headers for the given endpoint.

The default implementation returns an empty list.

RestClientScaffold.parameters — Function

parameters([config::RequestConfig], endpoint::AbstractEndpoint) -> Vector{Pair{String, String}}

Return URI parameters for the given endpoint.

This are combined with the endpoint URL to form the full query URL.

The default implementation returns an empty list.

Note

Part of the AbstractEndpoint interface.

RestClientScaffold.payload — Function

payload([config::RequestConfig], endpoint::AbstractEndpoint) -> Any

Return the payload for the given endpoint.

This is used for POST requests, and is sent as the body of the request.

Note

Part of the AbstractEndpoint interface.

Response interface

RestClientScaffold.validate — Function

validate([config::RequestConfig], endpoint::AbstractEndpoint) -> Bool

Check if the request to endpoint according to config is valid.

This is called before the request is made, and can be used to check if the request is well-formed. This is the appropriate place to emit warnings about potential issues with the request.

Return true if the request should proceed, false otherwise.

The default implementation always returns true.

Note

Part of the AbstractEndpoint interface.

RestClientScaffold.responsetype — Function

responsetype(endpoint::AbstractEndpoint) -> Type

Return the type of the response for the given endpoint.

Together with dataformat, this is used to parse the response.

If IO (the default implementation), the response is not parsed at all.

Note

Part of the AbstractEndpoint interface.

RestClientScaffold.postprocess — Function

postprocess([response::Downloads.Response], request::Request, data) -> Any

Post-process the data returned by the request.

There are three generic implementations provided:

For SingleEndpoint requests that return a SingleResponse, the data is wrapped in a Single object.

For ListEndpoint requests that return a ListResponse, the data are wrapped in a List object.

For all other endpoints, the data is returned as-is.

Note

Part of the AbstractEndpoint interface.

RestClientScaffold.contents — Function

content(response::SingleResponse{T}) -> T

Return the content of the response.

content(response::ListResponse{T}) -> Vector{T}

Return the items of the response.

RestClientScaffold.metadata — Function

metadata(response::SingleResponse) -> Dict{Symbol, Any}
metadata(response::ListResponse) -> Dict{Symbol, Any}

Return metadata for the given response.

The default implementation returns an empty dictionary.

RestClientScaffold.nextpage — Function

nextpage(response::List) -> Union{List, Nothing}

Fetch the next page of results after response.

If there are no more pages, or this method is not available for the given endpoint, return nothing.

RestClientScaffold.thispagenumber — Function

thispagenumber(response::List) -> Union{Int, Nothing}

Return the current page number of response, if known.

RestClientScaffold.remainingpages — Function

remainingpages(response::List) -> Union{Int, Nothing}

Return the number of remaining pages after response, if known.

Content formatting

RestClientScaffold.AbstractFormat — Type

AbstractFormat

Abstract supertype for response formats.

See also: RawFormat, JSONFormat.

RestClientScaffold.RawFormat — Type

RawFormat <: AbstractFormat

Singleton type for raw response formats.

RestClientScaffold.JSONFormat — Type

JSONFormat <: AbstractFormat

Singleton type for JSON response formats.

RestClientScaffold.dataformat — Function

dataformat([endpoint::AbstractEndpoint], ::Type{T}) -> AbstractFormat

Return the expected format that T is represented by in the response from endpoint.

Using the default dataformat(::Type) method, the format is RawFormat.

A dataformat(::Type) method is automatically defined when invoking @jsondef.

RestClientScaffold.interpretresponse — Function

interpretresponse(data::IO, fmt::AbstractFormat, ::Type{T}) -> value::T

Interpret data as a response of type T according to fmt.

RestClientScaffold.writepayload — Function

writepayload(dest::IO, fmt::AbstractFormat, data)

Write data to dest according to fmt.